Move getFileType into getCorrespondingType
