+++ src/utils/fileTypeToDisplay.js
@@ -1,0 +1,10 @@
+const fileTypeToDisplay = {
+  barcodes10x: 'barcodes.tsv',
+  features10x: 'genes.tsv',
+  matrix10x: 'matrix.mtx',
+  '10x_h5': 'matrix.h5',
+  seurat: 'seurat rds',
+  rhapsody: 'expression_data.st',
+};
+
+export default fileTypeToDisplay;
